and some more refactoring
