+++ src/components/app/app.tsx
@@ -1,6 +1,6 @@
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import MainScreen from '../../pages/main-screen/main-screen';
-import { AppRoute, Status } from '../utils/constants';
+import { AppRoute, Status } from '../../utils/constants';
 import FavoritesScreen from '../../pages/favorites/favorites-screen';
 import LoginScreen from '../../pages/login-screen/login-screen';
 import NotFoundScreen from '../../pages/not-found-screen/not-found-screen';
