+++ src/utils/formatRating.ts
